templates/requirements.txt
flask
python-docx
pandas
openpyxl
gunicorn